Extract S3BackingClient
